requirements.txt
boto3
requests
python-dotenv
slackclient